Divided view into more files, a few renamings
